requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
pyjwt>=2.9.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
